update: implement worst championList in summonerWinList component
