+++ src/app/page.tsx
@@ -8,7 +8,6 @@ import CardSection from "@/component/cardSection";
 import LargePlaceholderCard from "@/component/largePlaceholderCard";
 import RecentMatches from "@/component/recentMatches";
 import SummonerWinRateList from "@/component/summonerWinList";
-import SummonerWorstRateList from "@/component/summonerWorstList";
 import { mainGameName, mainNames, mainTagName } from "@/constant/basic";
 
 export default async function Home() {
@@ -113,16 +112,9 @@ export default async function Home() {
 				{/* 챔피언 승률 요약 */}
 				<div className="col-span-12">
 					<h1 className="text-2xl font-semibold text-gray-800 mb-2">
-						최근 {matchCount}게임 중 승률이 가장 높은 챔피언
+						최근 {matchCount}게임
 					</h1>
 					<SummonerWinRateList perSummonerStats={perSummonerStats} />
-				</div>
-
-				<div className="col-span-12">
-					<h1 className="text-2xl font-semibold text-gray-800 mb-2">
-						최근 {matchCount}게임 중 승률이 가장 낮은 챔피언
-					</h1>
-					<SummonerWorstRateList perSummonerStats={perSummonerStats} />
 				</div>
 
 				{/* 기타 카드 섹션 */}
