+++ src/component/summonerWinList.tsx
@@ -20,10 +20,17 @@ type TopChampion = {
 	rank: number;
 };
 
-type WinRateData = {
+type BottomChampion = {
+	name: string;
+	winRate: number;
+	rank: number;
+};
+
+type RateData = {
 	summoner: string;
 	champions: Champion[];
 	top3Champions: TopChampion[];
+	bottom3Champions: BottomChampion[];
 };
 
 // 순수 함수들
@@ -43,7 +50,19 @@ const createChampion = ([name, { wins, total }]: [
 const sortByWinRateDesc = (a: Champion, b: Champion): number =>
 	b.winRate - a.winRate;
 
+const sortByWinRateAsc = (a: Champion, b: Champion): number =>
+	a.winRate - b.winRate;
+
 const createTopChampion = (champion: Champion, index: number): TopChampion => ({
+	name: champion.name,
+	winRate: champion.winRate,
+	rank: index + 1,
+});
+
+const createBottomChampion = (
+	champion: Champion,
+	index: number,
+): BottomChampion => ({
 	name: champion.name,
 	winRate: champion.winRate,
 	rank: index + 1,
@@ -52,25 +71,31 @@ const createTopChampion = (champion: Champion, index: number): TopChampion => ({
 const getTop3Champions = (champions: Champion[]): TopChampion[] =>
 	champions.slice(0, 3).map(createTopChampion);
 
+const getBottom3Champions = (champions: Champion[]): BottomChampion[] =>
+	champions.slice(0, 3).map(createBottomChampion);
+
 const processStatsForSummoner = ([summoner, statsObj]: [
 	string,
 	Record<string, ChampionStats>,
-]): WinRateData => {
+]): RateData => {
 	const champions = Object.entries(statsObj)
 		.map(createChampion)
 		.sort(sortByWinRateDesc);
+
+	const getWorstChampions = (champions: Champion[]) =>
+		getBottom3Champions([...champions].sort(sortByWinRateAsc));
 
 	return {
 		summoner,
 		champions,
 		top3Champions: getTop3Champions(champions),
+		bottom3Champions: getWorstChampions(champions),
 	};
 };
 
 const transformPerSummonerStats = (
 	perSummonerStats: PerSummonerStats,
-): WinRateData[] =>
-	Object.entries(perSummonerStats).map(processStatsForSummoner);
+): RateData[] => Object.entries(perSummonerStats).map(processStatsForSummoner);
 
 // 스타일링 함수들
 const getRankColor = (rank: number): string => {
@@ -91,11 +116,36 @@ const getMedalIcon = (rank: number): string => {
 	return medalMap[rank] || "";
 };
 
+const getWorstIcon = (rank: number): string => {
+	const medalMap: Record<number, string> = {
+		1: "💀",
+		2: "😵",
+		3: "😰",
+	};
+	return medalMap[rank] || "";
+};
+
 const formatWinRate = (winRate: number): string =>
 	`${(winRate * 100).toFixed(1)}%`;
 
 const formatGameRecord = (wins: number, total: number): string =>
 	`(${wins}승 / ${total}패)`;
+
+const renderBottom3Item = ({ name, winRate, rank }: BottomChampion) => (
+	<div key={name} className="text-xs">
+		<span className="inline-block w-4">{rank}위</span>
+		<span className={`px-1 rounded ${getRankColor(rank)}`}>
+			{name}: {formatWinRate(winRate)}
+		</span>
+	</div>
+);
+
+const renderBottom3Section = (bottom3Champions: BottomChampion[]) => (
+	<div className="mb-3 p-2 bg-red-50 rounded">
+		<h4 className="text-sm font-medium text-red-800 mb-1">💔 Bottom 3</h4>
+		<div className="space-y-1">{bottom3Champions.map(renderBottom3Item)}</div>
+	</div>
+);
 
 // 컴포넌트 렌더링 함수들
 const renderTop3Item = ({ name, winRate, rank }: TopChampion) => (
@@ -155,13 +205,15 @@ const renderSummonerCard = ({
 	summoner,
 	champions,
 	top3Champions,
-}: WinRateData) => (
+	bottom3Champions,
+}: RateData) => (
 	<div
 		key={summoner}
 		className="flex-1 min-w-0 p-4 bg-white rounded-lg shadow-sm"
 	>
 		<h3 className="text-lg font-semibold mb-2">{summoner}님의 챔피언 승률</h3>
 		{renderTop3Section(top3Champions)}
+		{renderBottom3Section(bottom3Champions)}
 		{renderChampionsList(champions, top3Champions)}
 	</div>
 );
